double

+++ src/map.c
@@ -76,7 +76,7 @@ static void read_metadata(FILE *f, struct Map *map)
 		if (strstr(line, "Name=") == line)
 			snprintf(map->name, sizeof map->name, "%s", strstr(line, "=") + 1);
 		else if (strstr(line, "SortKey=") == line)
-			map->sortkey = (float)atof(strstr(line, "=") + 1);
+			map->sortkey = atof(strstr(line, "=") + 1);
 		else
 			log_printf_abort("bad metadata line: %s", line);
 	}
